Rework the whole thing to operate on a GaussianSplatAsset instead of direct ply file
The data is put into a bunch of textures, with intent that we
can change their formats sometime really soon, to take up
less space
