+++ Assets/GSTestScene.unity
@@ -302,8 +302,8 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1462254335}
   serializedVersion: 2
-  m_LocalRotation: {x: 0.90152323, y: 0.05462407, z: 0.42328975, w: 0.071399786}
-  m_LocalPosition: {x: -3.0026746, y: 1.4007691, z: 2.2284043}
+  m_LocalRotation: {x: 0.90152323, y: 0.054624073, z: 0.42328972, w: 0.07139974}
+  m_LocalPosition: {x: -3.0026817, y: 1.4007727, z: 2.2284005}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -338,8 +338,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: c12d929a7f62c48adbe9ff45c9a33ff0, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_PointCloudFolder: Assets/Models~/bicycle
-  m_Use30kVersion: 1
+  m_Asset: {fileID: 11400000, guid: 68d90f9ac82946a4aa799a15d76531de, type: 2}
   m_SplatScale: 1
   m_SHOrder: 3
   m_SortNthFrame: 1
@@ -347,7 +346,6 @@ MonoBehaviour:
   m_PointDisplaySize: 3
   m_DisplayData: 0
   m_PreferFfxSort: 1
-  m_ScaleDown: 1
   m_ShaderSplats: {fileID: 4800000, guid: ed800126ae8844a67aad1974ddddd59c, type: 3}
   m_ShaderComposite: {fileID: 4800000, guid: 7e184af7d01193a408eb916d8acafff9, type: 3}
   m_ShaderDebugPoints: {fileID: 4800000, guid: b44409fc67214394f8f47e4e2648425e, type: 3}
