+++ Assets/GaussianSplatting/Scripts/Editor/GaussianSplatAssetEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 75971a29a6deda14c9b1ff5f4ab2f2a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
