+++ Assets/GaussianSplatting/Shaders/GaussianDebugDisplayData.shader
@@ -17,24 +17,10 @@ CGPROGRAM
 
 #include "GaussianSplatting.hlsl"
 
-struct InputSplat
-{
-    float3 pos;
-    float3 nor;
-    float3 sh0;
-    float3 sh1, sh2, sh3, sh4, sh5, sh6, sh7, sh8, sh9, sh10, sh11, sh12, sh13, sh14, sh15;
-    float opacity;
-    float3 scale;
-    float4 rot;
-};
-StructuredBuffer<InputSplat> _DataBuffer;
-
 struct v2f
 {
     float4 vertex : SV_POSITION;
 };
-
-float _SplatSize;
 
 v2f vert (uint vtxID : SV_VertexID)
 {
@@ -50,6 +36,28 @@ v2f vert (uint vtxID : SV_VertexID)
     return o;
 }
 
+/*
+Morton interleaving 16x16 group i.e. by 4 bits of coordinates, based on https://twitter.com/rygorous/status/986715358852608000
+thread which is simplified version of https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/
+
+uint EncodeMorton2D_16x16(uint2 c)
+{
+    uint t = ((c.y & 0xF) << 8) | (c.x & 0xF); // ----EFGH----ABCD
+    t = (t ^ (t << 2)) & 0x3333;               // --EF--GH--AB--CD
+    t = (t ^ (t << 1)) & 0x5555;               // -E-F-G-H-A-B-C-D
+    return (t | (t >> 7)) & 0xFF;              // --------EAFBGCHD
+}
+
+uint2 DecodeMorton2D_16x16(uint t)
+{
+    t = (t & 0xFF) | ((t & 0xFE) << 7); // -EAFBGCHEAFBGCHD
+    t &= 0x5555;                        // -E-F-G-H-A-B-C-D
+    t = (t ^ (t >> 1)) & 0x3333;        // --EF--GH--AB--CD
+    t = (t ^ (t >> 2)) & 0x0f0f;        // ----EFGH----ABCD
+    return uint2(t & 0xF, t >> 8);
+}
+*/
+
 uint _SplatCount;
 uint _DisplayMode;
 
@@ -62,8 +70,6 @@ static const uint kDisplayRotation = 3;
 static const uint kDisplayColor = 4;
 static const uint kDisplayOpacity = 5;
 static const uint kDisplaySH1 = 6;
-
-static const float SH_C0 = 0.2820948;
 
 half4 frag (v2f i) : SV_Target
 {
@@ -81,7 +87,7 @@ half4 frag (v2f i) : SV_Target
         return 0;
 
     half4 res = 1;
-    InputSplat splat = _DataBuffer[idx];
+    SplatData splat = LoadSplatData(idx);
     if (_DisplayMode == kDisplayPosition)
     {
         float3 pos = (splat.pos * float3(1,1,-1) - _BoundsMin) / (_BoundsMax - _BoundsMin);
@@ -89,37 +95,35 @@ half4 frag (v2f i) : SV_Target
     }
     if (_DisplayMode == kDisplayRotation)
     {
-        float4 rot = normalize(splat.rot.yzwx);
-        res.rgb = saturate(rot.rgb * 0.5 + 0.5);
+        res.rgb = saturate(splat.rot.rgb * 0.5 + 0.5);
     }
     if (_DisplayMode == kDisplayScale)
     {
-        float3 scl = abs(exp(splat.scale));
-        res.rgb = saturate(scl * 3.0);
+        res.rgb = saturate(splat.scale * 3.0);
     }
     if (_DisplayMode == kDisplayColor)
     {
-        res.rgb = saturate((SH_C0 * splat.sh0 + 0.5) * 0.7);
+        res.rgb = saturate(splat.sh.col * 0.7);
     }
     if (_DisplayMode == kDisplayOpacity)
     {
-        res.rgb = saturate(Sigmoid(splat.opacity));
+        res.rgb = saturate(splat.opacity);
     }
-    if (_DisplayMode == kDisplaySH1 + 0) res.rgb = saturate(2 * splat.sh1 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 1) res.rgb = saturate(2 * splat.sh2 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 2) res.rgb = saturate(2 * splat.sh3 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 3) res.rgb = saturate(2 * splat.sh4 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 4) res.rgb = saturate(2 * splat.sh5 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 5) res.rgb = saturate(2 * splat.sh6 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 6) res.rgb = saturate(2 * splat.sh7 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 7) res.rgb = saturate(2 * splat.sh8 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 8) res.rgb = saturate(2 * splat.sh9 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 9) res.rgb = saturate(2 * splat.sh10 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 10) res.rgb = saturate(2 * splat.sh11 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 11) res.rgb = saturate(2 * splat.sh12 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 12) res.rgb = saturate(2 * splat.sh13 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 13) res.rgb = saturate(2 * splat.sh14 * 0.5 + 0.5);
-    if (_DisplayMode == kDisplaySH1 + 14) res.rgb = saturate(2 * splat.sh15 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 0) res.rgb = saturate(2 * splat.sh.sh1 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 1) res.rgb = saturate(2 * splat.sh.sh2 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 2) res.rgb = saturate(2 * splat.sh.sh3 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 3) res.rgb = saturate(2 * splat.sh.sh4 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 4) res.rgb = saturate(2 * splat.sh.sh5 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 5) res.rgb = saturate(2 * splat.sh.sh6 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 6) res.rgb = saturate(2 * splat.sh.sh7 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 7) res.rgb = saturate(2 * splat.sh.sh8 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 8) res.rgb = saturate(2 * splat.sh.sh9 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 9) res.rgb = saturate(2 * splat.sh.sh10 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 10) res.rgb = saturate(2 * splat.sh.sh11 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 11) res.rgb = saturate(2 * splat.sh.sh12 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 12) res.rgb = saturate(2 * splat.sh.sh13 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 13) res.rgb = saturate(2 * splat.sh.sh14 * 0.5 + 0.5);
+    if (_DisplayMode == kDisplaySH1 + 14) res.rgb = saturate(2 * splat.sh.sh15 * 0.5 + 0.5);
     return res;
 }
 ENDCG
