+++ Assets/GaussianSplatting/Shaders/GaussianDebugRenderBoxes.shader
@@ -29,19 +29,9 @@ static const int kCubeIndices[36] =
 #include "UnityCG.cginc"
 #include "GaussianSplatting.hlsl"
 
-struct InputSplat
-{
-    float3 pos;
-    float3 nor;
-    float3 sh0;
-    float3 sh1, sh2, sh3, sh4, sh5, sh6, sh7, sh8, sh9, sh10, sh11, sh12, sh13, sh14, sh15;
-    float opacity;
-    float3 scale;
-    float4 rot;
-};
-StructuredBuffer<InputSplat> _DataBuffer;
 StructuredBuffer<uint> _OrderBuffer;
 
+/*
 struct ChunkData
 {
     float3 bmin;
@@ -50,14 +40,13 @@ struct ChunkData
 StructuredBuffer<ChunkData> _ChunkBuffer;
 bool _DisplayChunks;
 int _ChunkCount;
+*/
 
 struct v2f
 {
     half4 col : COLOR0;
     float4 vertex : SV_POSITION;
 };
-
-static const float SH_C0 = 0.2820948;
 
 float _SplatScale;
 
@@ -71,7 +60,7 @@ half3 palette(float t, half3 a, half3 b, half3 c, half3 d)
 v2f vert (uint vtxID : SV_VertexID, uint instID : SV_InstanceID)
 {
     v2f o;
-    bool chunks = _DisplayChunks;
+    bool chunks = false; //@TODO _DisplayChunks;
 	uint idx = kCubeIndices[vtxID];
 	float3 localPos = float3(idx&1, (idx>>1)&1, (idx>>2)&1) * 2.0 - 1.0;
 
@@ -81,30 +70,32 @@ v2f vert (uint vtxID : SV_VertexID, uint instID : SV_InstanceID)
     {
         // display splat boxes
         instID = _OrderBuffer[instID];
-        InputSplat splat = _DataBuffer[instID];
+        SplatData splat = LoadSplatData(instID);
 
-        float4 boxRot = normalize(splat.rot.yzwx);
-        float3 boxSize = exp(splat.scale);
+        float4 boxRot = splat.rot;
+        float3 boxSize = splat.scale;
         boxSize *= _SplatScale;
 
         float3x3 splatRotScaleMat = CalcMatrixFromRotationScale(boxRot, boxSize);
 
         centerWorldPos = splat.pos * float3(1,1,-1);
 
-        o.col.rgb = saturate(SH_C0 * splat.sh0 + 0.5);
-        o.col.a = saturate(Sigmoid(splat.opacity));
+        o.col.rgb = saturate(splat.sh.col);
+        o.col.a = saturate(splat.opacity);
 
         localPos = mul(splatRotScaleMat, localPos) * 2;
     }
     else
     {
-        // display chunk boxes
+        // display chunk boxes @TODO
+        /*
         localPos = localPos * 0.5 + 0.5;
         ChunkData chunk = _ChunkBuffer[instID];
         localPos = lerp(chunk.bmin, chunk.bmax, localPos);
 
         o.col.rgb = palette((float)instID / (float)_ChunkCount, half3(0.5,0.5,0.5), half3(0.5,0.5,0.5), half3(1,1,1), half3(0.0, 0.33, 0.67));
         o.col.a = 0.1;
+        */
     }
     localPos.z *= -1;
